Updated files
Finished coding for launch page, still need to tweak the colors, but it
works. Also finished the coding for the gallery page, utilized a
Fisher-Yates shuffle algorithm for the randomization of images, and
chose to splice the imageFiles array every time the random button is
clicked. The coding all WORKS, but the images need to be tweaked with a
bit.